fix: purge counter overflow, registry null validation, benchmark thread safety
- Use long for totalDeleted in OutboxPurgeScheduler.runOnce() to prevent int overflow
- Add null checks for aggregateType and eventType in DefaultListenerRegistry.register()
- Add @Threads(1) to OutboxDispatchBenchmark to prevent latchRef race condition

+++ outbox-core/src/main/java/outbox/purge/OutboxPurgeScheduler.java
@@ -93,7 +93,7 @@ public final class OutboxPurgeScheduler implements AutoCloseable {
   public void runOnce() {
     try {
       Instant cutoff = Instant.now().minus(retention);
-      int totalDeleted = 0;
+      long totalDeleted = 0;
       int deleted;
       do {
         deleted = purgeBatch(cutoff);
